Add string to float mappings functions
The new flag will allow the user to use a YAML file which contains a
path with a set of mappings to convert strings to a float64 which is
able to be exported into Prometheus.

Take for example BGP session status which is sent as an integer with SNMP.

We could use a YAML file with the following structure to convert the
various states to the corresponding integers.

---
"state/session-state":
   IDLE: 1
   CONNECT: 2
   OPEN_SENT: 3  #Not yet sure about this one
   OPEN_CONFIRM: 4  #Not yet sure about this one
   ACTIVE: 5
   ESTABLISHED: 6

This would change the following:
key: state/session-state
str_value: ESTABLISHED

into

key: state/session-state
uint_value: 6