Fix to stop bad key/value pairs from being added.
Previously the switch/case included a continue statement for whenever
the error value of a string conversion was not nil. To add the string
mapping function the if statement had to become if/if else to validate
whether the key/value pair was part of one of the scenarios. I
accidentally removed the continue statement for when the error was not
nil. Fix was simply to an a closing else statement to continue if either
of two if cases were not matched.